Added support for sparse rules in files
This allows the sparse rules to be reused if a repository has multiple workflows and/or actions performing checkouts.

Added a new action input named sparse-checkout-file.

This setting overrides sparse-checkout in the event that both have been set.

If set then attempt to retrieve that file from the repository using git show (since the repository isn't checked out at this point).

The action will fail if it's not able to retrieve the sparse checkout rules file from the repository.

Added disabling of sparse checkout rules in the event that none are specified. This is important for self hosted runners that could be running workflows on the same repository that enable and disable the sparse checkout rules between commits.

The sparse checkout operations are now performed prior to the lfsFetch operation. This is to cover the scenario where the sparse rules were removed and the checkout isn't relied upon to fetch the objects lazily.